MODE-1713 - Transformed DocumentChanges and inner classes into interfaces, together with some other cleanup based on the code review
